requirements.txt
requests==2.31.0
clabe==1.2.4
lxml==4.9.1
dataclasses>=0.6;python_version<"3.7"
pytest==6.2.5
pytest-vcr==1.0.2
pytest-cov==3.0.0
black==22.3.0
flake8==4.0.1
isort[pipfile]==5.10.1
mypy==0.790
Flask==
mysql-connector-python
